refactor(NEAT): split NeatEvolution.ts into focused modules
Extract captureUtilisationSnapshot + computeOverallCpuUtilisation into a
new src/NEAT/CpuUtilisation.ts and processCompletedResults into a new
src/NEAT/ProcessCompletedResults.ts so NeatEvolution.ts is dominated by
the evolve() orchestrator instead of a ~280-line drain routine and a
pair of utilisation helpers. No behavioural change; existing evolve
tests pass.

Adds test/NEAT/CpuUtilisation.ts covering the now-independently-testable
utilisation helpers.

NeatEvolution.ts: 1141 -> 868 lines.

Refs #2395
